+++ stdnet/contrib/monitor/__init__.py
@@ -1,0 +1,5 @@
+'''
+A djpcms_ application for display admin panels of stdnet models
+
+.. _djpcms: https://github.com/lsbardel/djpcms
+'''